Add files via upload
Forage_task-1:
I was asked to record and submit an audio recording regarding ICO (Information Commissioner Officer) Dawn raids. I conducted research and analyzed what ICO dawn raids are, who may be liable of being subjected to a dawn raid, powers of an ICO, Procedures to be followed and precautions to be taken when a dawn raid occurs .

Forage_task-2:
I was asked to study a real life data exposure threat of a video game company and verify if complies with GDPR regulations and notification obligations 

Forage_task-3:
In this task I was asked to prepare a defensive strategy for a company Shikari Inc., which holds a data warehouse (cloud servers, etc) , was subjected to data breach due to a Cyber Attack and its client Ludens Inc., claims a notice saying their loss has to be paid. I studied GDPR regulations and gave a possible plan to defend against the litigations using possible potential arguments